+++ src/lib/breadcrumbs.module.ts
@@ -1,10 +1,10 @@
 import { NgModule } from '@angular/core';
-import { HubBreadcrumbsComponent } from './components/breadcrumbs/breadcrumbs.component';
+import { HubBreadcrumbComponent } from './components/breadcrumb/breadcrumb.component';
 import { HubBreadcrumbItemDirective } from './directives/breadcrumb-item.directive';
 
 @NgModule({
 	declarations: [],
-	imports: [HubBreadcrumbsComponent, HubBreadcrumbItemDirective],
-	exports: [HubBreadcrumbsComponent, HubBreadcrumbItemDirective]
+	imports: [HubBreadcrumbComponent, HubBreadcrumbItemDirective],
+	exports: [HubBreadcrumbComponent, HubBreadcrumbItemDirective]
 })
 export class HubBreadcrumbsModule {}