+++ src/lib/components/breadcrumb/breadcrumb.component.spec.ts
@@ -1,0 +1,24 @@
+import { ComponentFixture, TestBed } from '@angular/core/testing';
+import { provideRouter } from '@angular/router';
+import { HubBreadcrumbComponent } from './breadcrumb.component';
+
+describe('HubBreadcrumbComponent', () => {
+  let component: HubBreadcrumbComponent;
+  let fixture: ComponentFixture<HubBreadcrumbComponent>;
+
+  beforeEach(async () => {
+    await TestBed.configureTestingModule({
+      imports: [HubBreadcrumbComponent],
+      providers: [provideRouter([])]
+    })
+    .compileComponents();
+    
+    fixture = TestBed.createComponent(HubBreadcrumbComponent);
+    component = fixture.componentInstance;
+    fixture.detectChanges();
+  });
+
+  it('should create', () => {
+    expect(component).toBeTruthy();
+  });
+});