+++ src/lib/services/breadcrumbs.service.spec.ts
@@ -1,4 +1,5 @@
 import { TestBed } from '@angular/core/testing';
+import { provideRouter } from '@angular/router';
 
 import { HubBreadcrumbsService } from './breadcrumbs.service';
 
@@ -6,7 +7,9 @@ describe('HubBreadcrumbsService', () => {
 	let service: HubBreadcrumbsService;
 
 	beforeEach(() => {
-		TestBed.configureTestingModule({});
+		TestBed.configureTestingModule({
+			providers: [provideRouter([])]
+		});
 		service = TestBed.inject(HubBreadcrumbsService);
 	});