+++ src/public-api.ts
@@ -2,6 +2,6 @@
  * Public API Surface of breadcrumbs
  */
 export * from './lib/breadcrumbs.module';
-export * from './lib/components/breadcrumbs/breadcrumbs.component';
+export * from './lib/components/breadcrumb/breadcrumb.component';
 export * from './lib/directives/breadcrumb-item.directive';
 export * from './lib/services/breadcrumbs.service';