+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 All notable changes to this project will be documented in this file.
 
+## [21.1.0] - 2026-03-17
+
+### Changed
+
+- Refactored internal component structure: renamed `hub-breadcrumbs` to `hub-breadcrumb`.
+- Improved style encapsulation: manual style import is no longer required as styles are now bundled within the component.
+
 ## [21.0.0] - 2026-03-09
 
 ### Added